refactor(cloud): rename storage.py to azure.py
Names the module after the service it checks rather than the generic
storage concept, consistent with aws.py naming.
