+++ tools/cloud/__init__.py
@@ -6,8 +6,8 @@ These are the checks that should never be missed on any target.
 """
 
 from tools.cloud.aws import check_s3_buckets
+from tools.cloud.azure import check_azure_storage
 from tools.cloud.services import check_exposed_services
-from tools.cloud.storage import check_azure_storage
 
 __all__ = [
     "check_azure_storage",
