+++ tools/cloud/azure.py
